refactor: organize manager providers by version
