+++ projects/ngx-meta/src/core/src/managers/provider/v1/index.ts
@@ -1,0 +1,4 @@
+export {
+  makeMetadataManagerProviderFromSetterFactory,
+  MakeMetadataManagerProviderFromSetterFactoryOptions,
+} from './make-metadata-manager-provider-from-setter-factory'
