+++ projects/ngx-meta/src/core/src/managers/provider/v2/index.ts
@@ -1,0 +1,15 @@
+export {
+  provideNgxMetaManager,
+  withManagerDeps,
+  withManagerGlobal,
+  withManagerObjectMerging,
+  withManagerJsonPath,
+  _ProvideNgxMetaManagerOptions,
+} from './provide-ngx-meta-manager'
+export {
+  _provideNgxMetaModuleManager,
+  _ProvideNgxMetaModuleManagerOptions,
+  _withModuleManagerNameAttribute,
+  _withModuleManagerSetterFactory,
+  _withModuleManagerSameGlobalKey,
+} from './provide-ngx-meta-module-manager'
